pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "notebooklm-mcp-server"
version = "0.1.0"
description = "Connect AI LAB Jay AI Mentor - MCP Server for Google NotebookLM"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Connect AI Lab", email = "jay@connexionai.kr"},
]
dependencies = [
    "mcp>=0.1.0",
    "httpx>=0.27.0",
    "browser_cookie3",
    "beautifulsoup4",
    "pydantic",
    "notebooklm-py" 
]

[project.scripts]
notebooklm-mcp-server = "notebooklm_mcp.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/notebooklm_mcp"]
